requirements.txt
Django==2.2.10
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-polymorphic==2.1.2
djangorestframework==3.9.4
factory-boy==2.12.0
Faker==4.1.2
iso639==0.1.4
Markdown==3.1.1
mysql==0.0.2
mysqlclient==1.4.2.post1
pbr==5.2.1
Pillow==6.2.0
python-dateutil==2.8.1
pytz==2019.1
PyYAML==5.1
six==1.12.0
sqlparse==0.3.0
stevedore==1.30.1
text-unidecode==1.3
virtualenv==16.6.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
